requirements.txt
dj-rest-auth==2.1.4
Django==2.2.6
django-rest-auth==0.9.5
djangorestframework==3.10.3
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
